Unified eureka security configuration
