+++ discovery-service/build.gradle
@@ -32,7 +32,6 @@ gitProperties {
 }
 
 dependencies {
-    compile(project(':apiml-common'))
     compile(project(':security-service-client-spring'))
 
     compile libraries.spring_boot_starter_web
